+++ src/metrics.rs
@@ -30,6 +30,8 @@ pub(crate) struct Metrics {
 
     pub(crate) seasonal_backgrounds: IntCounter,
 
+    pub(crate) score: IntCounter,
+
     pub(crate) own_data: IntCounter,
     pub(crate) user: IntCounter,
     pub(crate) user_beatmapsets: IntCounter,
@@ -75,6 +77,8 @@ impl Metrics {
             spotlights: counters.with_label_values(&["Spotlights"]),
 
             seasonal_backgrounds: counters.with_label_values(&["Seasonal backgrounds"]),
+
+            score: counters.with_label_values(&["Score"]),
 
             own_data: counters.with_label_values(&["Own Data"]),
             user: counters.with_label_values(&["User"]),